DEP-3909: Created CommonPastChatSurvey to bring in functionality for PostChatSurveyWebchat
