+++ app/assets/javascripts/views/postChatSurvey/PostChatSurveyDigitalAssistant.js
@@ -1,4 +1,4 @@
-import PrintUtils from "../../utils/PrintUtils"
+import CommonPostChatSurvey from "./CommonPostChatSurvey"
 
 const html = `
 <div id="postChatSurvey">
@@ -137,82 +137,12 @@ const html = `
 </div>
 `
 
-export default class PostChatSurveyDigitalAssistant {
-    constructor(onSubmitted) {
-        this.onSubmitted = onSubmitted;
-    }
-
-    attachTo(container) {
-        this.container = container;
-
-        this.wrapper = document.createElement("div");
-        this.wrapper.id = "postChatSurveyWrapper";
-        this.wrapper.insertAdjacentHTML("beforeend", html);
-        container.appendChild(this.wrapper);
-
-        this.wrapper.querySelector("#submitPostChatSurvey").addEventListener(
-            "click",
-            (e) => {
-                this.onSubmitted(this);
-            }
-        );
-
-        $('input[name="q3-"]').on('click', function() {
-           if ($(this).val() != 'Other') {
-                document.getElementById("q4-").value = "";
-           }
-        });
-
-        this.wrapper.querySelector("#printPostChat").addEventListener(
-          "click",
-          (e) => {
-              e.preventDefault;
-              this.onPrintPostChatSurvey(this);
-          }
-      );
-
-      window.addEventListener('afterprint', (event) => {
-        this.showTranscriptAndSurvey(false, true)
-      });
-
-    }
-
-    showTranscriptAndSurvey(showTranscript, showSurvey) {
-      let transcript = document.getElementById("ciapiSkinChatTranscript");
-      transcript.style.display = showTranscript ? "" : "none";
-
-      let postChatSurvey = document.getElementById("postChatSurveyWrapper");
-      postChatSurvey.style.display = showSurvey ? "" : "none";
-    }
-
-    onPrintPostChatSurvey(e) {
-      e.preventDefault;
-
-      this.showTranscriptAndSurvey(true, false);
-
-      document.getElementById("print-date").innerHTML = PrintUtils.getPrintDate();
-
-      const elementList = [
-          "app-related-items",
-          "govuk-back-link",
-          "govuk-phase-banner",
-          "hmrc-report-technical-issue",
-          "govuk-footer",
-          "govuk-heading-xl",
-          "hmrc-user-research-banner",
-          "cbanner-govuk-cookie-banner",
-          "postChatSurveyWrapper"
-      ];
-
-      PrintUtils.removeElementsForPrint(elementList);
-
-      window.print();
-
-      return false;
+export default class PostChatSurveyDigitalAssistant extends CommonPostChatSurvey {
+  constructor(onSubmitted) {
+      super(html, onSubmitted)
   }
 
-
-    detach() {
-        this.container.removeChild(this.wrapper)
-    }
+  // attachTo(container) {
+  //   this.CommonPostChatSurvey.attachTo()
+  // }
 }
